+++ CHANGELOG.md
@@ -16,6 +16,8 @@ This is a feature release.
    `nil` slice in `AdminClient.ListConsumerGroupOffsets`, when earlier it
    was not processing that correctly (#985, @alexandredantas).
  * Deprecate m.LeaderEpoch in favor of m.TopicPartition.LeaderEpoch (#1012).
+ * [KIP-554](https://cwiki.apache.org/confluence/display/KAFKA/KIP-554%3A+Add+Broker-side+SCRAM+Config+API):
+   User SASL/SCRAM credentials alteration and description (#1004).
 
 confluent-kafka-go is based on librdkafka v2.2.0, see the
 [librdkafka v2.2.0 release notes](https://github.com/confluentinc/librdkafka/releases/tag/v2.2.0-RC2)